Move compiler_stub to intrinsiclib
Fix the issue: #1354
