+++ unit_test/test_size/test_size_of_spdm_responder/CMakeLists.txt
@@ -2,14 +2,17 @@ cmake_minimum_required(VERSION 2.8.12)
 
 if(CMAKE_SYSTEM_NAME MATCHES "Linux")
     if(ARCH STREQUAL "x64")
-        SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint")
+        SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__stack_chk_fail")
+        if(TOOLCHAIN STREQUAL "CLANG")
+            SET(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS}  -Wl,--undefined=memcpy")
+        endif()
     elseif(ARCH STREQUAL "aarch64")
         if(TOOLCHAIN STREQUAL "ARM_DS2022")
-            SET(CMAKE_EXE_LINKER_FLAGS "--target=aarch64-arm-none-eabi -nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__udivti3")
+            SET(CMAKE_EXE_LINKER_FLAGS "--target=aarch64-arm-none-eabi -nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__udivti3,--undefined=memcpy")
         elseif(TOOLCHAIN STREQUAL "ARM_GNU")
-            SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__udivti3")
+            SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__udivti3,--undefined=memcpy")
         elseif(TOOLCHAIN STREQUAL "AARCH64_GCC")
-            SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint")
+            SET(CMAKE_EXE_LINKER_FLAGS "-nostdlib -flto -Wl,--entry,ModuleEntryPoint,--undefined=__stack_chk_fail,--undefined=__stack_chk_guard,--undefined=memcpy")
         endif()
     endif()
 elseif(CMAKE_SYSTEM_NAME MATCHES "Windows")
@@ -25,7 +28,6 @@ INCLUDE_DIRECTORIES(${LIBSPDM_DIR}/unit_test/test_size/test_size_of_spdm_respond
 SET(src_test_size_of_spdm_responder
     spdm_responder_init.c
     spdm_responder_main.c
-    compiler_stub.c
     support.c
 )
 
@@ -78,5 +80,3 @@ else()
     ADD_EXECUTABLE(test_size_of_spdm_responder ${src_test_size_of_spdm_responder})
     TARGET_LINK_LIBRARIES(test_size_of_spdm_responder ${test_size_of_spdm_responder_LIBRARY})
 endif()
-
-
